refactor: directory명 변경 (schema -> entity)
