+++ backend/src/main/java/capstone/backend/domain/auth/entity/RefreshToken.java
@@ -1,4 +1,4 @@
-package capstone.backend.domain.auth.schema;
+package capstone.backend.domain.auth.entity;
 
 import java.util.concurrent.TimeUnit;
