+++ backend/src/main/java/capstone/backend/domain/auth/repository/BlacklistTokenRedisRepository.java
@@ -1,6 +1,6 @@
 package capstone.backend.domain.auth.repository;
 
-import capstone.backend.domain.auth.schema.BlacklistToken;
+import capstone.backend.domain.auth.entity.BlacklistToken;
 import org.springframework.data.keyvalue.repository.KeyValueRepository;
 
 public interface BlacklistTokenRedisRepository extends KeyValueRepository<BlacklistToken, String> {
