+++ backend/src/main/java/capstone/backend/domain/auth/service/AuthService.java
@@ -8,9 +8,9 @@ import capstone.backend.domain.auth.exception.TokenNotFoundException;
 import capstone.backend.domain.auth.repository.BlacklistTokenRedisRepository;
 import capstone.backend.domain.auth.repository.OauthCodeRedisRepository;
 import capstone.backend.domain.auth.repository.RefreshTokenRedisRepository;
-import capstone.backend.domain.auth.schema.BlacklistToken;
-import capstone.backend.domain.auth.schema.OauthCode;
-import capstone.backend.domain.auth.schema.RefreshToken;
+import capstone.backend.domain.auth.entity.BlacklistToken;
+import capstone.backend.domain.auth.entity.OauthCode;
+import capstone.backend.domain.auth.entity.RefreshToken;
 import capstone.backend.global.security.jwt.JwtProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
