+++ backend/src/main/java/capstone/backend/domain/bubble/controller/v2/BubbleController.java
@@ -8,7 +8,7 @@ import capstone.backend.domain.bubble.service.BubbleService;
 import capstone.backend.domain.bubble.service.WebFluxService;
 import capstone.backend.domain.eisenhower.dto.request.EisenhowerItemCreateRequest;
 import capstone.backend.domain.inventory.dto.request.InventoryItemCreateRequest;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import capstone.backend.domain.member.service.MemberService;
 import capstone.backend.global.api.dto.ApiResponse;
 import capstone.backend.global.security.oauth2.user.CustomOAuth2User;
