+++ backend/src/main/java/capstone/backend/domain/eisenhower/entity/EisenhowerItem.java
@@ -2,7 +2,7 @@ package capstone.backend.domain.eisenhower.entity;
 
 import capstone.backend.domain.eisenhower.dto.request.EisenhowerItemCreateRequest;
 import capstone.backend.domain.eisenhower.dto.request.EisenhowerItemUpdateRequest;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EntityListeners;
