+++ backend/src/main/java/capstone/backend/domain/eisenhower/service/EisenhowerCategoryService.java
@@ -10,7 +10,7 @@ import capstone.backend.domain.eisenhower.repository.EisenhowerCategoryRepositor
 import capstone.backend.domain.eisenhower.repository.EisenhowerItemRepository;
 import capstone.backend.domain.member.exception.MemberNotFoundException;
 import capstone.backend.domain.member.repository.MemberRepository;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
