+++ backend/src/main/java/capstone/backend/domain/inventory/service/InventoryFolderService.java
@@ -8,7 +8,7 @@ import capstone.backend.domain.inventory.dto.request.InventoryFolderRequest;
 import capstone.backend.domain.inventory.dto.response.InventoryFolderResponse;
 import capstone.backend.domain.member.exception.MemberNotFoundException;
 import capstone.backend.domain.member.repository.MemberRepository;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
