+++ backend/src/main/java/capstone/backend/domain/member/entity/Role.java
@@ -1,4 +1,4 @@
-package capstone.backend.domain.member.scheme;
+package capstone.backend.domain.member.entity;
 
 import lombok.Getter;
