+++ backend/src/main/java/capstone/backend/domain/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package capstone.backend.domain.member.repository;
 
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
