+++ backend/src/main/java/capstone/backend/domain/member/service/MemberService.java
@@ -3,7 +3,7 @@ package capstone.backend.domain.member.service;
 import capstone.backend.domain.member.dto.response.UserDataDTO;
 import capstone.backend.domain.member.exception.MemberNotFoundException;
 import capstone.backend.domain.member.repository.MemberRepository;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
