+++ backend/src/main/java/capstone/backend/domain/mindmap/entity/MindMap.java
@@ -1,6 +1,6 @@
 package capstone.backend.domain.mindmap.entity;
 
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import capstone.backend.domain.common.entity.TaskType;
 import capstone.backend.domain.mindmap.dto.request.MindMapRequest;
 import capstone.backend.domain.mindmap.dto.request.UpdateMindMapRequest;
