+++ backend/src/main/java/capstone/backend/domain/pomodoro/entity/Pomodoro.java
@@ -1,6 +1,6 @@
-package capstone.backend.domain.pomodoro.schema;
+package capstone.backend.domain.pomodoro.entity;
 
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import capstone.backend.domain.pomodoro.exception.PomodoroDurationException;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
