+++ backend/src/main/java/capstone/backend/domain/pomodoro/service/PomodoroService.java
@@ -11,8 +11,8 @@ import capstone.backend.domain.pomodoro.dto.response.SidebarResponse;
 import capstone.backend.domain.pomodoro.exception.PomodoroDurationException;
 import capstone.backend.domain.pomodoro.exception.PomodoroNotFoundException;
 import capstone.backend.domain.pomodoro.repository.PomodoroRepository;
-import capstone.backend.domain.pomodoro.schema.Pomodoro;
-import capstone.backend.domain.pomodoro.schema.PomodoroCycle;
+import capstone.backend.domain.pomodoro.entity.Pomodoro;
+import capstone.backend.domain.pomodoro.entity.PomodoroCycle;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
