+++ backend/src/main/java/capstone/backend/global/security/jwt/JwtAuthenticationFilter.java
@@ -3,11 +3,10 @@ package capstone.backend.global.security.jwt;
 import capstone.backend.domain.auth.exception.AccessLogoutTokenException;
 import capstone.backend.domain.auth.exception.AccessTokenExpiredException;
 import capstone.backend.domain.auth.exception.InvalidTokenException;
-import capstone.backend.domain.auth.exception.TokenNotFoundException;
 import capstone.backend.domain.auth.repository.BlacklistTokenRedisRepository;
 import capstone.backend.domain.member.exception.MemberNotFoundException;
 import capstone.backend.domain.member.repository.MemberRepository;
-import capstone.backend.domain.member.scheme.Member;
+import capstone.backend.domain.member.entity.Member;
 import capstone.backend.global.api.dto.ApiResponse;
 import capstone.backend.global.api.exception.ApiException;
 import capstone.backend.global.security.oauth2.user.CustomOAuth2User;
